# Add benchmarks examining the performance difference between explicitly closing buffers, and letting Cleaners close the buffers