@normanmaurer 2020-11-25
The some super interesting observation.... Like you said this needs some further investigation. That said I think there is no way to go mixed mode here, either buffers should be closed explicit in general or not. Otherwise it will be mostly impossible for people to ensure the code is right etc.