@chrisvest 2020-11-25
I agree, we should go one way or the other. It's hard to influence the GC behaviour, but we can try to make `send()` cheaper.